requirements.txt
pyexiftool
typer
sh
